requirements.txt
robotframework
robotframework-seleniumlibrary
robotframework-metrics
selenium 
setuptools
virtualenv
